+++ GSC.Data/Entities/LabManagement/LabManagementVariableMapping.cs
@@ -15,10 +15,10 @@ namespace GSC.Data.Entities.LabManagement
         public int? AuditReasonId { get; set; }
         public AuditReason AuditReason { get; set; }
         public string ReasonOth { get; set; }
-        public int? MaleLowRange { get; set; }
-        public int? MaleHighRange { get; set; }
-        public int? FemaleLowRange { get; set; }
-        public int? FemaleHighRange { get; set; }
+        public decimal? MaleLowRange { get; set; }
+        public decimal? MaleHighRange { get; set; }
+        public decimal? FemaleLowRange { get; set; }
+        public decimal? FemaleHighRange { get; set; }
         public string Unit { get; set; }
     }
 }
